% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{ab_boundary}
\alias{ab_boundary}
\title{Alberta provincial boundary}
\usage{
ab_boundary()
}
\value{
A \code{SpatVector} containing one polygon.
}
\description{
Returns the 2020 revision of the Alberta provincial boundary as a
single dissolved polygon in \code{\link[=ab_crs]{ab_crs()}}.  This is the default
boundary used by \code{\link[=mask_to_boundary]{mask_to_boundary()}}.
}
\examples{
\dontrun{
ab <- ab_boundary()
terra::ext(ab)
}

}
\seealso{
\code{\link[=mask_to_boundary]{mask_to_boundary()}} to clip a raster to this boundary.
}
