% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{print.sciSpatial_metadata}
\alias{print.sciSpatial_metadata}
\title{Print a metadata object}
\usage{
\method{print}{sciSpatial_metadata}(x, ...)
}
\arguments{
\item{x}{A \code{sciSpatial_metadata} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a metadata object
}
