% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{print.sciSpatial_variables}
\alias{print.sciSpatial_variables}
\title{Print a compact view of the variables table}
\usage{
\method{print}{sciSpatial_variables}(x, ...)
}
\arguments{
\item{x}{A \code{sciSpatial_variables} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a compact view of the variables table
}
