% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_to_grid.R
\name{aggregate_to_grid}
\alias{aggregate_to_grid}
\title{Coarsen a raster to a reference grid}
\usage{
aggregate_to_grid(x, ref = ab_grid(), fun = "mean", categorical = FALSE, ...)
}
\arguments{
\item{x}{A \code{SpatRaster} to aggregate.}

\item{ref}{A \code{SpatRaster} defining the target (coarser) grid.
Defaults to the ABMI 1 km Alberta grid, \code{\link[=ab_grid]{ab_grid()}}.}

\item{fun}{Character or function; aggregation function for
continuous layers.  One of \code{"mean"}, \code{"sum"}, \code{"max"},
\code{"min"}, or any function accepted by \code{\link[terra:aggregate]{terra::aggregate()}}.
Ignored when \code{categorical = TRUE}.  Default \code{"mean"}.}

\item{categorical}{Logical; if \code{TRUE}, forces \code{fun = "modal"}.
Default \code{FALSE}.}

\item{...}{Additional arguments passed to \code{\link[terra:aggregate]{terra::aggregate()}}.}
}
\value{
A \code{SpatRaster} with approximately the same resolution as
\code{ref}.
}
\description{
Aggregates \code{x} to the (coarser) resolution of \code{ref} using an
integer factor derived from the resolution ratio.  For continuous
layers the aggregation function defaults to the mean; for
categorical layers it is forced to \code{\link[terra:modal]{terra::modal()}}.
}
\examples{
\dontrun{
library(terra)
ref <- rast(res = 1000, crs = "EPSG:3400",
            xmin = 0, xmax = 10000,
            ymin = 0, ymax = 10000)
r   <- rast(res = 250,  crs = "EPSG:3400",
            xmin = 0, xmax = 10000,
            ymin = 0, ymax = 10000)
r[] <- runif(ncell(r))
r_agg <- aggregate_to_grid(r, ref)

# Onto the default ABMI 1 km Alberta grid
r_agg2 <- aggregate_to_grid(r)
}

}
